// apps/api/src/businesses/dto/create-business.dto.d.ts
export declare class CreateBusinessDto {
    name: string;
    ein?: string;
    mobile?: string;
    country?: string;
    businessType?: string;
    industry?: string;
    employeeCount?: string;
    address?: string;
    city?: string;
    state?: string;
    zip?: string;
    modules?: string;
}