// apps/api/src/businesses/dto/update-business.dto.d.ts
export declare class UpdateBusinessDto {
    name?: string;
    ein?: string;
    mobile?: string;
    country?: string;
    businessType?: string;
    industry?: string;
    employeeCount?: string;
    currencyCode?: string;
    address?: string;
    city?: string;
    state?: string;
    zip?: string;
    modules?: string;
    logoUrl?: string;
    settings?: string;
}